requirements.txt
python-telegram-bot==21.4
fastapi==0.115.0
uvicorn==0.30.6
reportlab==4.2.2